[CHANGED] Moved particle pipeline file back to pipelines folder
